+++ packages/azure-func/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2023-09-20
+
+### Added
+
+- Function generator added.
+
+### Changed
+
+- Application generator now includes a sample spec file.
+
 ## [0.3.1] - 2023-09-17
 
 ### Fixed
@@ -25,7 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `serve` now watches the project for changes.
 
-[unreleased]: https://github.com/ziacik/nx-tools/compare/azure-func-0.3.1...HEAD
+[unreleased]: https://github.com/ziacik/nx-tools/compare/azure-func-1.0.0...HEAD
+[1.0.0]: https://github.com/ziacik/nx-tools/compare/azure-func-1.0.0...azure-func-0.3.1
 [0.3.1]: https://github.com/ziacik/nx-tools/compare/azure-func-0.3.1...azure-func-0.3.0
 [0.3.0]: https://github.com/ziacik/nx-tools/compare/azure-func-0.3.0...azure-func-0.2.0
 [0.2.0]: https://github.com/ziacik/nx-tools/releases/tag/azure-func-0.2.0